finish

+++ libft/ft_putstr_fd.c
@@ -6,13 +6,13 @@
 /*   By: jeounpar <jeounpar@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/08 00:41:35 by jeounpar          #+#    #+#             */
-/*   Updated: 2021/12/21 15:29:54 by jeounpar         ###   ########.fr       */
+/*   Updated: 2021/12/21 23:47:44 by jeounpar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t ft_putstr_fd(char *s, int fd)
+int	ft_putstr_fd(char *s, int fd)
 {
 	int	i;
